+++ depynd/bayesian_networks/__init__.py
@@ -1,4 +1,4 @@
-from depynd.bayesian_networks.sparse_candidate_ import mmpc
+from depynd.bayesian_networks.mmpc_ import mmpc
 from depynd.bayesian_networks.mmhc_ import mmhc
 
 __all__ = ['mmpc', 'mmhc']
